removing stream stuff away!

+++ 31_final_E_json/bus_catalog.h
@@ -15,7 +15,7 @@
 
 #include "json.h"
 
-using RealDistance = std::pair<std::string_view, size_t>;
+using RealDistance = std::pair<std::string_view, int>;
 
 std::ostream& operator<<(std::ostream& out, const RealDistance& el);
 
@@ -41,7 +41,7 @@ struct BusInfo{
     std::list<std::string_view> stop_names_;
     std::unordered_set<std::string_view> unique_stops_ = {};
     double route_len_ = 0.0;
-    size_t real_dist_ = 0;
+    int real_dist_ = 0;
 
     BusInfo(std::string_view id, Type type, std::list<std::string_view> snames);
 };
@@ -60,7 +60,7 @@ bool operator==(const std::pair<std::string_view, std::string_view>& lhs, const 
 class BusCatalog{
     using StopStruct = std::unordered_map<std::string_view, std::unique_ptr<Stop>>;
     using BusStruct = std::unordered_map<std::string_view, std::unique_ptr<BusInfo>>;
-    using DistStruct = std::unordered_map<std::pair<std::string_view, std::string_view>, size_t, PairHasher>;
+    using DistStruct = std::unordered_map<std::pair<std::string_view, std::string_view>, int, PairHasher>;
 
 
 private:
@@ -73,12 +73,11 @@ private:
     BusStruct buses_;
     DistStruct all_distances_;
 
-
     std::string_view SaveStopName(std::string_view new_stop_name);
     std::string_view SaveBusName(std::string_view new_bus_name);
     double ComputeDistance(std::string_view lhs, std::string_view rhs) const;
     double ComputeBusRouteLen(const BusInfo* bus) const;
-    size_t ComputeBusRealDist(const BusInfo* bus) const;
+    int ComputeBusRealDist(const BusInfo* bus) const;
     std::unordered_set<std::string_view> GetUniqueStops(const BusInfo* bus) const;
     void UpdateBusRouteLens();
     void UpdateBusUniqueStops();
@@ -93,7 +92,7 @@ public:
     std::optional<std::set<std::string_view>> GetBusesForStop(std::string_view stop_key) const ;
     const Stop* GetStop(std::string_view stop_key) const ;
     const BusInfo* GetBus(std::string_view bus_id) const ;
-    size_t GetDistStructSize() const ;
+    int GetDistStructSize() const ;
 
 };
